Fraternity elections are pretty boring to sit through so I decided to work on a rough draft of feature 12
Going to figure out if product owner wants to be able to set the
required values with a ArrayList or HashSet. Regardless of what product
owners wants to use I can always just convert a ArrayList to a HashSet
so that we can get constant time operation when see if a command line
argument is contained in a HashSet.