Big upload ***read comments for status update***
Couple of things:
Dr. Garret wants us to separate out our test files. Right now we have
one big test files called ArgParserTest.java, but instead we should
have a test file for each of our classes i.e. ArgTest.java,
OptArgTest.java and so forth. It will make the file sizes smaller and
cleaner. I think Arg and Opt arg should be done. We should finish up
the one for the XML file and and then go through the ArgParseTest (when
we get time) and take out some of the redundant tests.

Second I think with the feature we pulled so far we are almost done
besides a couple of things. One fixing the implementation so flags can
have descriptions because right now they don’t and then adjusting the
help message accordingly.

Additionally separating the way we create the XML from the ArgParser
class so it makes it easier to change code without breaking stuff and
cleaner.

 I created a acceptance test for feature 08 (just so I have experience
creating one) so don’t worry about if they don’t pass. I think i messed
up in creating two different error messages in a robot file.

I think features wise we should be good, so the next step will be
getting a good start on the documentation and figuring out ways to
refractor the code to make it cleaner and easier on the product owner’s
behalf to use.